requirements.txt
# Core dependencies for DictationApp
faster-whisper>=1.0.0
pyaudio>=0.2.14
pyperclip>=1.8.0

# Optional: AI refinement providers
# Uncomment if you want to use Gemini for text refinement
# google-generativeai>=0.3.0

# Remote transcription
google-genai>=1.0.0
